+++ NEWS.md
@@ -13,6 +13,8 @@ Bugfix
 
 2. Not ignoring 'allowRegularization' in calendar time settings.
 
+3. When person has multiple nesting cohort entries, prevent double-counting of cases in attrition table if nesting cohort entries have no outcome.
+
 
 SelfControlledCaseSeries 4.1.0
 ==============================